Add field validation
