+++ met-web/src/components/common/Input/FormField.tsx
@@ -1,16 +1,28 @@
 import React from 'react';
 import { Grid, GridProps } from '@mui/material';
 import { BodyText } from '../Typography';
+import { colors } from '..';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faExclamationCircle } from '@fortawesome/pro-regular-svg-icons';
 
 export type FormFieldProps = {
     title: string;
     instructions?: string;
     required?: boolean;
     optional?: boolean;
+    error?: string;
     children: React.ReactNode;
 } & GridProps;
 
-export const FormField = ({ title, instructions, required, optional, children, ...gridProps }: FormFieldProps) => {
+export const FormField = ({
+    title,
+    instructions,
+    required,
+    optional,
+    error,
+    children,
+    ...gridProps
+}: FormFieldProps) => {
     return (
         <Grid className="met-input-form-field" container spacing={0} direction="column" {...gridProps}>
             <Grid item xs={12}>
@@ -23,6 +35,14 @@ export const FormField = ({ title, instructions, required, optional, children, .
             <Grid item xs={12} sx={{ mb: '8px' }}>
                 <BodyText size="small">{instructions}</BodyText>
             </Grid>
+            {error && (
+                <Grid item xs={12} sx={{ mb: '8px' }}>
+                    <BodyText bold size="small" sx={{ color: colors.notification.error.shade }}>
+                        <FontAwesomeIcon icon={faExclamationCircle} style={{ marginRight: '8px', fontSize: '18px' }} />
+                        {error}
+                    </BodyText>
+                </Grid>
+            )}
             <Grid item container xs={12}>
                 {children}
             </Grid>
