+++ met-web/src/components/common/Input/TextInput.tsx
@@ -1,28 +1,30 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { Button as MuiButton, ButtonProps as MuiButtonProps, Input, InputProps, Box } from '@mui/material';
 import { colors, elevations, globalFocusVisible } from '..';
 import { FormField, FormFieldProps } from './FormField';
 import { BodyText } from '../Typography';
 import { faCircleXmark } from '@fortawesome/pro-regular-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { error } from 'console';
 
 type TextInputProps = {
     value?: string;
     onChange?: (value: string) => void;
     placeholder?: string;
     disabled?: boolean;
-    error?: string;
-} & Omit<InputProps, 'value' | 'onChange' | 'placeholder' | 'disabled' | 'error'>;
+} & Omit<InputProps, 'value' | 'onChange' | 'placeholder' | 'disabled'>;
 
 export const TextInput: React.FC<TextInputProps> = ({
     value,
     onChange,
     placeholder,
     disabled,
+    sx,
+    error,
+    inputProps,
     ...textFieldProps
 }: TextInputProps) => {
-    // Exclude props with conflicting types
-    const { error, sx, inputProps, ...rest } = textFieldProps;
+    // Exclude props that are not meant for the input element
     return (
         <Input
             className="met-input-text"
@@ -32,7 +34,6 @@ export const TextInput: React.FC<TextInputProps> = ({
             onChange={(e) => onChange?.(e.target.value)}
             placeholder={placeholder}
             disabled={disabled}
-            error={!!error}
             sx={{
                 display: 'flex',
                 height: '48px',
@@ -42,8 +43,7 @@ export const TextInput: React.FC<TextInputProps> = ({
                 gap: '10px',
                 alignSelf: 'stretch',
                 borderRadius: '8px',
-                outline: `1px solid ${colors.surface.gray[80]}`,
-                outlineStyle: 'inset',
+                outline: error ? `2px solid #B80C0D` : `1px solid ${colors.surface.gray[80]}`,
                 caretColor: colors.surface.blue[90],
                 '&:hover': {
                     outline: `2px solid ${colors.surface.gray[90]}`,
@@ -55,6 +55,7 @@ export const TextInput: React.FC<TextInputProps> = ({
                 ...sx,
             }}
             inputProps={{
+                error: error,
                 ...inputProps,
                 sx: {
                     fontSize: '16px',
@@ -67,55 +68,63 @@ export const TextInput: React.FC<TextInputProps> = ({
                     ...inputProps?.sx,
                 },
             }}
-            {...rest}
+            {...textFieldProps}
         />
     );
 };
 
 export type TextFieldProps = {
+    error?: string;
     counter?: boolean;
     maxLength?: number;
     clearable?: boolean;
 } & Omit<FormFieldProps, 'children'> &
-    Omit<TextInputProps, 'fullWidth'>;
+    Omit<TextInputProps, 'fullWidth' | 'error'>;
 
 export const TextField = ({
     title,
     instructions,
+    error,
     required,
     optional,
     clearable,
+    onChange,
     ...textInputProps
 }: TextFieldProps) => {
     const [value, setValue] = React.useState(textInputProps.value || '');
+    useEffect(() => {
+        setValue(textInputProps.value || '');
+    }, [textInputProps.value]);
     const handleSetValue = (newValue: string) => {
-        setValue(newValue);
-        textInputProps.onChange?.(newValue);
+        onChange?.(newValue) ?? setValue(newValue);
     };
-    const endAdornment = clearable
-        ? {
-              endAdornment: (
-                  <MuiButton
-                      onClick={() => handleSetValue('')}
-                      sx={{
-                          padding: '0',
-                          minWidth: 'unset',
-                          minHeight: 'unset',
-                          width: '24px',
-                          height: '24px',
-                          borderRadius: '50%',
-                          background: 'none',
-                          color: colors.type.regular.secondary,
-                          '&:hover': {
-                              background: colors.surface.gray[80],
-                          },
-                      }}
-                  >
-                      <FontAwesomeIcon icon={faCircleXmark} />
-                  </MuiButton>
-              ),
-          }
-        : {};
+    const isError = !!error;
+    const endAdornment =
+        clearable && value.length > 0
+            ? {
+                  endAdornment: (
+                      <MuiButton
+                          onClick={() => handleSetValue('')}
+                          title="Clear this field"
+                          sx={{
+                              minWidth: 'unset',
+                              minHeight: 'unset',
+                              width: '24px',
+                              height: '24px',
+                              borderRadius: '50%',
+                              background: 'none',
+                              color: colors.surface.gray[70],
+                              '&:hover, &:focus': {
+                                  color: colors.surface.gray[90],
+                              },
+                              ...globalFocusVisible,
+                          }}
+                      >
+                          <FontAwesomeIcon icon={faCircleXmark} />
+                      </MuiButton>
+                  ),
+              }
+            : {};
     const length = value.length;
     return (
         <FormField
@@ -124,8 +133,18 @@ export const TextField = ({
             instructions={instructions}
             required={required}
             optional={optional}
+            error={error}
         >
-            <TextInput fullWidth value={value} {...endAdornment} {...textInputProps} onChange={handleSetValue} />
+            <TextInput
+                fullWidth
+                error={isError}
+                value={value}
+                required={required}
+                {...endAdornment}
+                {...textInputProps}
+                inputProps={{ ...textInputProps.inputProps, maxLength: textInputProps.maxLength }}
+                onChange={handleSetValue}
+            />
             <Box sx={{ width: '100%', display: 'flex', justifyContent: 'flex-end', gap: '8px' }}>
                 {textInputProps.counter && textInputProps.maxLength && (
                     <BodyText size="small" sx={{ color: colors.type.regular.secondary }}>
