+++ met-web/src/components/common/index.tsx
@@ -109,6 +109,24 @@ export const colors = {
         },
         white: '#FFFFFF',
     },
+    notification: {
+        success: {
+            shade: '#42814A',
+            tint: '#F6FFF8',
+        },
+        warning: {
+            shade: '#FCBA19',
+            tint: '#FFECBE',
+        },
+        danger: {
+            shade: '#CE3E39',
+            tint: '#F4E1E2',
+        },
+        error: {
+            shade: '#CE3E39',
+            tint: '#F4E1E2',
+        },
+    },
 };
 
 export const elevations = {
